M Start of authentication

+++ restart.sh
@@ -1,7 +1,9 @@
-killall node
-node /opt/lysanoe/current/src/index.js &
-exit
-
+if [ -z $1 ]; then 
+	echo '--- Restarting NodeJS only'	
+	killall node
+	node src/index.js &
+	exit;
+fi
 
 ./stop.sh
 sleep 1
